Time: 72 ms (19.96%), Space: 59.3 MB (45.76%) - LeetHub

+++ 2665-minimum-time-to-repair-cars/2665-minimum-time-to-repair-cars.cpp
@@ -13,11 +13,11 @@ public:
         return ans;
     }
 
-    long long repairCars(vector<int>& ranks, long long cars) {
+    long long repairCars(vector<int>& ranks, int cars) {
         
         int n = ranks.size();
         long long l = 1;
-        long long r = 100 * cars * cars;
+        long long r = 100LL * cars * cars;
         
         long long ans = r;
         while( l <= r)
